Move code into its own file
